Ensure upper value of an int range is an int
While it is stored as a double precision float, an int should be exactly
returned up to huge values.
https://stackoverflow.com/questions/1848700/biggest-integer-that-can-be-stored-in-a-double

+++ application/libraries/Attribute_Value_ORM.php
@@ -64,7 +64,7 @@ abstract class Attribute_Value_ORM extends ORM {
           $vf = 'int_value';
           $array->add_rules('int_value', 'integer');
           if ($attr->allow_ranges === 't') {
-            $array->add_rules('upper_value', 'numeric');
+            $array->add_rules('upper_value', 'integer');
           }
           break;
